requirements.txt
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.0
h11==0.16.0
httptools==0.7.1
idna==3.11
limits==5.8.0
packaging==26.0
passlib==1.7.4
psutil==7.2.2
pyasn1==0.6.2
pycparser==3.0
pydantic==2.7.4
pydantic-settings==2.2.1
pydantic_core==2.18.4
psycopg[binary]==3.2.10
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.44
starlette==0.36.3
typing_extensions==4.15.0
uvicorn==0.29.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
python-json-logger
transformers==4.57.6
tokenizers==0.22.1
safetensors==0.6.2
torch==2.8.0
youtube-transcript-api==1.2.4
